Save page rank in database. sdaDocument add, remove through MongoStore

+++ A1/src/crawler/src/main/java/edu/carleton/comp4601/CrawlerVertex.java
@@ -10,17 +10,17 @@ public class CrawlerVertex implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	// Mirrors structure of document storage in Mongo
-	private long ID;
+	private int ID;
 	private String URL;
 	
-	public CrawlerVertex(long ID, String URL) {
+	public CrawlerVertex(int ID, String URL) {
 		this.ID = ID;
 		this.URL = URL;
 	}
 	
 	public CrawlerVertex(Page page) {
 		this(
-			(long) page.getWebURL().getDocid(),
+			(int) page.getWebURL().getDocid(),
 			page.getWebURL().toString()
 		);
 	}
@@ -30,9 +30,9 @@ public class CrawlerVertex implements Serializable {
 		return URL.substring(URL.lastIndexOf("/"));
 	}
 
-	public long getID() { return ID; }
+	public int getID() { return ID; }
 
-	public void setID(long iD) { ID = iD; }
+	public void setID(int iD) { ID = iD; }
 
 	public String getURL() { return URL; }
